Module rusoto_signature::signature

AWS API request signatures.

Follows AWS Signature 4 algorithm.

If needed, the request will be re-issued to a temporary redirect endpoint. This can happen with newly created S3 buckets not in us-standard/us-east-1.

Please note that this module does not expect URIs to already be encoded.

Structs

SignedRequest
A data structure for all the elements of an HTTP request that are involved in the Amazon Signature Version 4 signing process

Enums

SignedRequestPayload
Possible payloads included in a SignedRequest.

Constants

STRICT_ENCODE_SET
This constant is used to maintain the strict URI encoding standard as proposed by RFC 3986

STRICT_PATH_ENCODE_SET
This struct is used to maintain the URI path encoding

Statics

EMPTY_SHA256_HASH
Payload string to use for signed empty payload

UNSIGNED_PAYLOAD
Payload string to use for unsigned payload

Functions

string_to_sign
Mark string as AWS4-HMAC-SHA256 hashed

Type Definitions

Params